+++ chromecast/media/cma/backend/mixer/mixer_loopback_connection.h
@@ -48,7 +48,7 @@ class MixerLoopbackConnection : public mixer_service::MixerSocket::Delegate {
  private:
   // mixer_service::MixerSocket::Delegate implementation:
   bool HandleMetadata(const mixer_service::Generic& message) override;
-  bool HandleAudioData(char* data, int size, int64_t timestamp) override;
+  bool HandleAudioData(char* data, size_t size, int64_t timestamp) override;
   void OnConnectionError() override;
 
   const std::unique_ptr<mixer_service::MixerSocket> socket_;
